webapp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.49.0
pyyaml>=6.0
python-frontmatter>=1.1.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
playwright>=1.48.0
aiofiles>=24.1.0
